requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
langchain-google-genai==0.0.10
pydantic==2.6.3
python-dotenv==1.0.1